+++ src/DependencyInjection/WebgriffeSyliusAkeneoExtension.php
@@ -47,7 +47,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
             'arguments' => [
                 'sylius.factory.product_image',
                 'sylius.repository.product_image',
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
             ],
         ],
         'immutable_slug' => [
@@ -62,7 +62,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'product_option' => [
             'class' => ProductOptionValueHandler::class,
             'arguments' => [
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
                 'sylius.repository.product_option',
                 'sylius.factory.product_option_value',
                 'sylius.factory.product_option_value_translation',
@@ -89,7 +89,7 @@ final class WebgriffeSyliusAkeneoExtension extends AbstractResourceExtension imp
         'file_attribute' => [
             'class' => FileAttributeValueHandler::class,
             'arguments' => [
-                'webgriffe_sylius_akeneo.api_client_official',
+                'webgriffe_sylius_akeneo.api_client',
                 'filesystem',
             ],
         ],
