+++ src/Resources/config/services.xml
@@ -19,27 +19,17 @@
         </service>
 
         <!-- General -->
-        <service id="webgriffe_sylius_akeneo.api_client" class="Webgriffe\SyliusAkeneoPlugin\ApiClient" >
-            <argument type="service" id="sylius.http_client"/>
-            <argument>%webgriffe_sylius_akeneo.api_client.base_url%</argument>
-            <argument>%webgriffe_sylius_akeneo.api_client.username%</argument>
-            <argument>%webgriffe_sylius_akeneo.api_client.password%</argument>
-            <argument>%webgriffe_sylius_akeneo.api_client.client_id%</argument>
-            <argument>%webgriffe_sylius_akeneo.api_client.secret%</argument>
-        </service>
-        <service id="Webgriffe\SyliusAkeneoPlugin\ApiClientInterface" alias="webgriffe_sylius_akeneo.api_client" />
-
-        <service id="webgriffe_sylius_akeneo.api_client_official.factory" class="Akeneo\Pim\ApiClient\AkeneoPimClientBuilder">
+        <service id="webgriffe_sylius_akeneo.api_client.factory" class="Akeneo\Pim\ApiClient\AkeneoPimClientBuilder">
             <argument>%webgriffe_sylius_akeneo.api_client.base_url%</argument>
         </service>
-        <service id="webgriffe_sylius_akeneo.api_client_official" class="Akeneo\Pim\ApiClient\AkeneoPimClient">
-            <factory service="webgriffe_sylius_akeneo.api_client_official.factory" method="buildAuthenticatedByPassword" />
+        <service id="webgriffe_sylius_akeneo.api_client" class="Akeneo\Pim\ApiClient\AkeneoPimClient">
+            <factory service="webgriffe_sylius_akeneo.api_client.factory" method="buildAuthenticatedByPassword" />
             <argument>%webgriffe_sylius_akeneo.api_client.client_id%</argument>
             <argument>%webgriffe_sylius_akeneo.api_client.secret%</argument>
             <argument>%webgriffe_sylius_akeneo.api_client.username%</argument>
             <argument>%webgriffe_sylius_akeneo.api_client.password%</argument>
         </service>
-        <service id="Akeneo\Pim\ApiClient\AkeneoPimClientInterface" alias="webgriffe_sylius_akeneo.api_client_official" />
+        <service id="Akeneo\Pim\ApiClient\AkeneoPimClientInterface" alias="webgriffe_sylius_akeneo.api_client" />
 
         <service id="webgriffe_sylius_akeneo.product.value_handlers_resolver" class="Webgriffe\SyliusAkeneoPlugin\PriorityValueHandlersResolver" />
         <service id="Webgriffe\SyliusAkeneoPlugin\ValueHandlersResolverInterface" alias="webgriffe_sylius_akeneo.product.value_handlers_resolver" />
@@ -60,7 +50,7 @@
         <service id="Webgriffe\SyliusAkeneoPlugin\Product\TaxonsResolverInterface" alias="webgriffe_sylius_akeneo.product.taxons_resolver" />
 
         <service id="webgriffe_sylius_akeneo.product.product_options_resolver" class="Webgriffe\SyliusAkeneoPlugin\Product\ProductOptionsResolver">
-            <argument type="service" id="webgriffe_sylius_akeneo.api_client_official"/>
+            <argument type="service" id="webgriffe_sylius_akeneo.api_client"/>
             <argument type="service" id="sylius.repository.product_option" />
             <argument type="service" id="sylius.factory.product_option" />
             <argument type="service" id="sylius.factory.product_option_translation" />
@@ -80,7 +70,7 @@
             <argument type="service" id="sylius.factory.product_variant" />
             <argument type="service" id="sylius.repository.product_variant" />
             <argument type="service" id="sylius.repository.product" />
-            <argument type="service" id="webgriffe_sylius_akeneo.api_client_official" />
+            <argument type="service" id="webgriffe_sylius_akeneo.api_client" />
             <argument type="service" id="webgriffe_sylius_akeneo.product.value_handlers_resolver" />
             <argument type="service" id="sylius.factory.product" />
             <argument type="service" id="webgriffe_sylius_akeneo.product.taxons_resolver" />
@@ -95,7 +85,7 @@
 
         <!-- Product Associations Importer -->
         <service id="webgriffe_sylius_akeneo.product_associations.importer" class="Webgriffe\SyliusAkeneoPlugin\ProductAssociations\Importer" >
-            <argument type="service" id="webgriffe_sylius_akeneo.api_client_official"/>
+            <argument type="service" id="webgriffe_sylius_akeneo.api_client"/>
             <argument type="service" id="sylius.repository.product"/>
             <argument type="service" id="sylius.repository.product_association"/>
             <argument type="service" id="sylius.repository.product_association_type"/>
