+++ tests/Application/config/services_test.yaml
@@ -9,8 +9,6 @@ services:
 
     webgriffe_sylius_akeneo.api_client:
         class: 'Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\ApiClientMock'
-    webgriffe_sylius_akeneo.api_client_official:
-        class: 'Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\OfficialApiClientMock'
     webgriffe_sylius_akeneo.date_time_builder:
         class: 'Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\DateTimeBuilder'
