+++ tests/Behat/Context/Setup/AkeneoContext.php
@@ -6,12 +6,12 @@ namespace Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Context\Setup;
 
 use Akeneo\Pim\ApiClient\AkeneoPimClientInterface;
 use Behat\Behat\Context\Context;
-use Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\OfficialApiClientMock;
+use Tests\Webgriffe\SyliusAkeneoPlugin\Integration\TestDouble\ApiClientMock;
 use Webmozart\Assert\Assert;
 
 final class AkeneoContext implements Context
 {
-    /** @var AkeneoPimClientInterface|OfficialApiClientMock */
+    /** @var AkeneoPimClientInterface|ApiClientMock */
     private $apiClient;
 
     public function __construct(AkeneoPimClientInterface $apiClient)
@@ -24,7 +24,7 @@ final class AkeneoContext implements Context
      */
     public function thereIsAProductUpdatedAtOnAkeneo(string $identifier, \DateTime $date)
     {
-        Assert::isInstanceOf($this->apiClient, OfficialApiClientMock::class);
+        Assert::isInstanceOf($this->apiClient, ApiClientMock::class);
         $this->apiClient->addProductUpdatedAt($identifier, $date);
     }
 
@@ -33,7 +33,7 @@ final class AkeneoContext implements Context
      */
     public function thereAreProductsOnAkeneo(int $count)
     {
-        Assert::isInstanceOf($this->apiClient, OfficialApiClientMock::class);
+        Assert::isInstanceOf($this->apiClient, ApiClientMock::class);
         for ($i = 1; $i <= $count; ++$i) {
             $this->apiClient->addProductUpdatedAt('product-' . $i, new \DateTime());
         }
